Isolate entry types
